Большая задача по Maps API. Часть №3

+++ Draw.py
@@ -1,9 +1,11 @@
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QPixmap
-from PyQt5.QtWidgets import QApplication, QMainWindow
+from PyQt5.QtWidgets import QMainWindow
 from MainWindow import Ui_MainWindow
 from RequestsSystem import RequestsSystem
 import os
+
+EPSILON = 0.001
 
 
 class MiniSearch(QMainWindow, Ui_MainWindow):
@@ -12,42 +14,67 @@ class MiniSearch(QMainWindow, Ui_MainWindow):
         self.setupUi(self)
         self.requestsSystem = RequestsSystem()
         self.setFixedSize(800, 600)
-        self.showMapbtn.clicked.connect(lambda: self.showMap(
-            [self.lon.text(), self.lat.text()], self.scale.text(), "map"))
+        self.showMapbtn.clicked.connect(lambda: self.showMap())
+        self.mapFile = "map.png"
+        self.currentViewMap = "map"
 
-    def showMap(self, coords, scale, viewMap="map", mark=""):
+    def loadMap(self):
+        with open(self.mapFile, "wb") as file:
+            self.respone = self.requestsSystem.getMap(
+                [self.currentLon, self.currentLat], self.currentScale,
+                self.currentViewMap, self.currentMark)
+            if self.respone.ok:
+                file.write(self.respone.content)
+            else:
+                self.labelError.setText("Ошибка запроса")
+        self.pixmap = QPixmap(self.mapFile)
+        self.photo.setPixmap(self.pixmap)
+
+    def updateData(self, mark=""):
+        self.currentLon = self.lon.text()
+        self.currentLat = self.lat.text()
+        self.currentScale = self.scale.text()
+        self.currentMark = mark
+
+    def clearErrors(self):
+        self.labelErrorData.setText("")
+        self.labelError.setText("")
+
+    def showMap(self, mark=""):
         # валидаторы
+        self.updateData()
         if not (self.lon.text() and self.lat.text() and self.scale.text()) != "":
             self.labelErrorData.setText("Введите данные")
         else:
-            self.currentCoords = coords
-            self.currentScale = scale
-            self.currentViewMap = viewMap
-            self.currentMark = mark
-            self.labelErrorData.setText("")
-            self.labelError.setText("")
-            self.mapFile = "map.png"
-            with open(self.mapFile, "wb") as file:
-                self.respone = self.requestsSystem.getMap(
-                    self.currentCoords, self.currentScale, self.currentViewMap, self.currentMark)
-                if self.respone.ok:
-                    file.write(self.respone.content)
-                else:
-                    self.labelError.setText("Ошибка запроса")
-            self.pixmap = QPixmap(self.mapFile)
-            self.photo.setPixmap(self.pixmap)
+            self.clearErrors()
+            self.loadMap()
 
     def keyPressEvent(self, event):
         if event.key() == Qt.Key_PageUp:
             if 0 <= int(self.scale.text()) < 17:
-                self.currentScale = str(int(self.currentScale) + 1)
-                self.scale.setText(self.currentScale)
-                self.showMap(self.currentCoords, self.currentScale, self.currentViewMap, self.currentMark)
+                self.scale.setText(str(int(self.currentScale) + 1))
+                self.showMap()
         if event.key() == Qt.Key_PageDown:
             if 0 <= int(self.scale.text()) <= 17:
-                self.currentScale = str(int(self.currentScale) - 1)
-                self.scale.setText(self.currentScale)
-                self.showMap(self.currentCoords, self.currentScale, self.currentViewMap, self.currentMark)
+                self.scale.setText(str(int(self.currentScale) - 1))
+                self.showMap()
+        if event.key() == Qt.Key_Up:
+            if abs(float(self.currentLon)) <= 180:
+                self.lat.setText(str(round(float(self.currentLat) + EPSILON, 6)))
+                self.showMap()
+        if event.key() == Qt.Key_Down:
+            if abs(float(self.currentLon)) <= 180:
+                self.lat.setText(str(round(float(self.currentLat) - EPSILON, 6)))
+                self.showMap()
+        if event.key() == Qt.Key_Left:
+            if abs(float(self.currentLon)) <= 180:
+                self.lon.setText(str(round(float(self.currentLon) - EPSILON, 6)))
+                self.showMap()
+        if event.key() == Qt.Key_Right:
+            if abs(float(self.currentLon)) <= 180:
+                self.lon.setText(str(round(float(self.currentLon) + EPSILON, 6)))
+                self.showMap()
 
     def closeEvent(self, event):
-        os.remove(self.mapFile)
+        if os.path.exists("map.png"):
+            os.remove(self.mapFile)
